requirements.txt
Pillow==12.1.1
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
numpy
Pillow
